requirements.txt
Flask
opencv-python
numpy
scikit-learn
requests
python-dotenv
